refactor: update telegram send message structure
