+++ README.md
@@ -3,11 +3,11 @@
 Add `Telegram` protocol to Spaceone's Notification channel! This plugin lets you integrate telegram-bot service and spaceone's core notification service by using [telegram's bot API](https://core.telegram.org/bots/api). <br>
 
 ## How to Setup plugin
-You can use SpaceONE's command line tool [Spacectl](https://github.com/spaceone-dev/spacectl) to setup and initialize plugin. 
-[Here](https://github.com/spaceone-dev/plugin-telegram-noti-protocol/tree/master/test/spacectl) shows the examples of manifest files to use telegram-notification plugin.
+You can use SpaceONE's command line tool [Spacectl](https://github.com/cloudforet-io/spacectl) to setup and initialize plugin. 
+[Here](https://github.com/cloudforet-io/plugin-telegram-noti-protocol/tree/master/test/spacectl) shows the examples of manifest files to use telegram-notification plugin.
 
 ## How to Setup Telegram notification channel in SpaceONE portal
-[user guide](https://www.spaceone.org/docs/guides/user_guide/notification/protocol/telegram/) will help you. 
+[user guide](docs/ko/GUIDE.md) will help you. 
 
 ### For more information..
-Visit our [website](https://www.spaceone.org/).
+Visit our [website](https://cloudforet.io/).
